refactor(frontend): extract ChartSkeleton + candlestick data helpers, dedup RS strip layout
